chore: archive Mars rover photos grid view and switch to timeline display
