+++ src/app/archived/mars-rover-photos-v2/mars-rover-photos.module.css
@@ -1,0 +1,32 @@
+.mainContainer {
+  margin-top: 4rem;
+}
+
+.heading {
+  font-size: 1.5rem;
+  margin: 0.5rem 1rem;
+  color: var(--secondary);
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: 1rem;
+  margin: 1rem;
+}
+
+.loading {
+  color: var(--info);
+  margin: 0.5rem 0;
+}
+
+@media (min-width: 768px) {
+  .mainContainer {
+    width: 768px;
+    margin: 4rem auto;
+  }
+
+  .grid {
+    grid-template-columns: 1fr 1fr;
+  }
+}
